+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/PlatformApiProxy.java
@@ -24,9 +24,9 @@ import com.refinedmods.refinedstorage2.platform.api.network.node.exporter.Export
 import com.refinedmods.refinedstorage2.platform.api.network.node.externalstorage.PlatformExternalStorageProviderFactory;
 import com.refinedmods.refinedstorage2.platform.api.network.node.importer.ImporterTransferStrategyFactory;
 import com.refinedmods.refinedstorage2.platform.api.registry.PlatformRegistry;
+import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ResourceFactory;
-import com.refinedmods.refinedstorage2.platform.api.resource.ResourceInstance;
 import com.refinedmods.refinedstorage2.platform.api.resource.ResourceRendering;
 import com.refinedmods.refinedstorage2.platform.api.storage.StorageRepository;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.PlatformStorageChannelType;
@@ -34,13 +34,12 @@ import com.refinedmods.refinedstorage2.platform.api.storage.type.StorageType;
 import com.refinedmods.refinedstorage2.platform.api.upgrade.UpgradeRegistry;
 
 import java.util.Collection;
-import java.util.Optional;
+import java.util.Set;
 import javax.annotation.Nullable;
 
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
 public class PlatformApiProxy implements PlatformApi {
@@ -207,6 +206,21 @@ public class PlatformApiProxy implements PlatformApi {
     }
 
     @Override
+    public ResourceFactory<ItemResource> getItemResourceFactory() {
+        return ensureLoaded().getItemResourceFactory();
+    }
+
+    @Override
+    public ResourceFactory<FluidResource> getFluidResourceFactory() {
+        return ensureLoaded().getFluidResourceFactory();
+    }
+
+    @Override
+    public Set<ResourceFactory<?>> getAlternativeResourceFactories() {
+        return ensureLoaded().getAlternativeResourceFactories();
+    }
+
+    @Override
     public <T> void registerResourceRendering(final Class<T> resourceClass, final ResourceRendering<T> rendering) {
         ensureLoaded().registerResourceRendering(resourceClass, rendering);
     }
@@ -214,11 +228,6 @@ public class PlatformApiProxy implements PlatformApi {
     @Override
     public <T> ResourceRendering<T> getResourceRendering(final T resource) {
         return ensureLoaded().getResourceRendering(resource);
-    }
-
-    @Override
-    public Optional<ResourceInstance<?>> createResource(final ItemStack stack, final boolean tryAlternatives) {
-        return ensureLoaded().createResource(stack, tryAlternatives);
     }
 
     @Override