+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/PlatformApiImpl.java
@@ -29,9 +29,9 @@ import com.refinedmods.refinedstorage2.platform.api.network.node.exporter.Export
 import com.refinedmods.refinedstorage2.platform.api.network.node.externalstorage.PlatformExternalStorageProviderFactory;
 import com.refinedmods.refinedstorage2.platform.api.network.node.importer.ImporterTransferStrategyFactory;
 import com.refinedmods.refinedstorage2.platform.api.registry.PlatformRegistry;
+import com.refinedmods.refinedstorage2.platform.api.resource.FluidResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
 import com.refinedmods.refinedstorage2.platform.api.resource.ResourceFactory;
-import com.refinedmods.refinedstorage2.platform.api.resource.ResourceInstance;
 import com.refinedmods.refinedstorage2.platform.api.resource.ResourceRendering;
 import com.refinedmods.refinedstorage2.platform.api.storage.StorageRepository;
 import com.refinedmods.refinedstorage2.platform.api.storage.channel.PlatformStorageChannelType;
@@ -46,6 +46,7 @@ import com.refinedmods.refinedstorage2.platform.common.internal.grid.PlatformGri
 import com.refinedmods.refinedstorage2.platform.common.internal.item.StorageContainerHelperImpl;
 import com.refinedmods.refinedstorage2.platform.common.internal.network.LevelConnectionProvider;
 import com.refinedmods.refinedstorage2.platform.common.internal.registry.PlatformRegistryImpl;
+import com.refinedmods.refinedstorage2.platform.common.internal.resource.FluidResourceFactory;
 import com.refinedmods.refinedstorage2.platform.common.internal.resource.ItemResourceFactory;
 import com.refinedmods.refinedstorage2.platform.common.internal.storage.ClientStorageRepository;
 import com.refinedmods.refinedstorage2.platform.common.internal.storage.StorageRepositoryImpl;
@@ -67,7 +68,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.PriorityQueue;
 import java.util.Queue;
 import java.util.Set;
@@ -77,7 +77,6 @@ import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
@@ -121,7 +120,8 @@ public class PlatformApiImpl implements PlatformApi {
     );
     private final List<GridExtractionStrategyFactory> gridExtractionStrategyFactories = new ArrayList<>();
     private final List<GridScrollingStrategyFactory> gridScrollingStrategyFactories = new ArrayList<>();
-    private final ResourceFactory itemResourceFactory = new ItemResourceFactory();
+    private final ResourceFactory<ItemResource> itemResourceFactory = new ItemResourceFactory();
+    private final ResourceFactory<FluidResource> fluidResourceFactory = new FluidResourceFactory();
     private final Set<ResourceFactory<?>> resourceFactories = new HashSet<>();
     private final Map<Class<?>, ResourceRendering<?>> resourceRenderingMap = new HashMap<>();
 
@@ -341,6 +341,21 @@ public class PlatformApiImpl implements PlatformApi {
     }
 
     @Override
+    public ResourceFactory<ItemResource> getItemResourceFactory() {
+        return itemResourceFactory;
+    }
+
+    @Override
+    public ResourceFactory<FluidResource> getFluidResourceFactory() {
+        return fluidResourceFactory;
+    }
+
+    @Override
+    public Set<ResourceFactory<?>> getAlternativeResourceFactories() {
+        return resourceFactories;
+    }
+
+    @Override
     public <T> void registerResourceRendering(final Class<T> resourceClass, final ResourceRendering<T> rendering) {
         resourceRenderingMap.put(resourceClass, rendering);
     }
@@ -349,20 +364,6 @@ public class PlatformApiImpl implements PlatformApi {
     @SuppressWarnings("unchecked")
     public <T> ResourceRendering<T> getResourceRendering(final T resource) {
         return (ResourceRendering<T>) resourceRenderingMap.get(resource.getClass());
-    }
-
-    @Override
-    @SuppressWarnings({"rawtypes", "unchecked"})
-    public Optional<ResourceInstance<?>> createResource(final ItemStack stack, final boolean tryAlternatives) {
-        if (tryAlternatives) {
-            for (final ResourceFactory factory : resourceFactories) {
-                final Optional<ResourceInstance<?>> resourceInstance = factory.create(stack);
-                if (resourceInstance.isPresent()) {
-                    return resourceInstance;
-                }
-            }
-        }
-        return itemResourceFactory.create(stack);
     }
 
     @Override