+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/block/FluidStorageBlockContainerMenu.java
@@ -1,9 +1,9 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu.storage.block;
 
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.storage.StorageConfigurationContainer;
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 import com.refinedmods.refinedstorage2.platform.common.internal.resource.ResourceContainer;
-import com.refinedmods.refinedstorage2.platform.common.internal.storage.channel.StorageChannelTypes;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
@@ -13,7 +13,13 @@ public class FluidStorageBlockContainerMenu extends AbstractStorageBlockContaine
     public FluidStorageBlockContainerMenu(final int syncId,
                                           final Inventory playerInventory,
                                           final FriendlyByteBuf buf) {
-        super(Menus.INSTANCE.getFluidStorage(), syncId, playerInventory.player, buf, StorageChannelTypes.FLUID);
+        super(
+            Menus.INSTANCE.getFluidStorage(),
+            syncId,
+            playerInventory.player,
+            buf,
+            PlatformApi.INSTANCE.getFluidResourceFactory()
+        );
     }
 
     public FluidStorageBlockContainerMenu(final int syncId,