+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/block/ItemStorageBlockContainerMenu.java
@@ -1,9 +1,9 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu.storage.block;
 
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.storage.StorageConfigurationContainer;
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 import com.refinedmods.refinedstorage2.platform.common.internal.resource.ResourceContainer;
-import com.refinedmods.refinedstorage2.platform.common.internal.storage.channel.StorageChannelTypes;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Inventory;
@@ -11,7 +11,13 @@ import net.minecraft.world.entity.player.Player;
 
 public class ItemStorageBlockContainerMenu extends AbstractStorageBlockContainerMenu {
     public ItemStorageBlockContainerMenu(final int syncId, final Inventory playerInventory, final FriendlyByteBuf buf) {
-        super(Menus.INSTANCE.getItemStorage(), syncId, playerInventory.player, buf, StorageChannelTypes.ITEM);
+        super(
+            Menus.INSTANCE.getItemStorage(),
+            syncId,
+            playerInventory.player,
+            buf,
+            PlatformApi.INSTANCE.getItemResourceFactory()
+        );
     }
 
     public ItemStorageBlockContainerMenu(final int syncId,