+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/screen/tooltip/MouseWithIconClientTooltipComponent.java
@@ -1,19 +1,13 @@
 package com.refinedmods.refinedstorage2.platform.common.screen.tooltip;
 
-import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
-import com.refinedmods.refinedstorage2.platform.api.resource.ItemResource;
 import com.refinedmods.refinedstorage2.platform.common.screen.TextureIds;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import javax.annotation.Nullable;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
-import net.minecraft.world.item.ItemStack;
 
 public class MouseWithIconClientTooltipComponent implements ClientTooltipComponent {
     private final Type type;
@@ -50,31 +44,6 @@ public class MouseWithIconClientTooltipComponent implements ClientTooltipCompone
             graphics.drawString(font, amount, x + 9 + 4 + 16 - font.width(amount), y + 9, 16777215, true);
             poseStack.popPose();
         }
-    }
-
-    public static List<ClientTooltipComponent> createForFilter(final ItemStack carried) {
-        if (carried.isEmpty()) {
-            return Collections.emptyList();
-        }
-        final List<ClientTooltipComponent> lines = new ArrayList<>();
-        PlatformApi.INSTANCE.createResource(carried, false).ifPresent(asItem -> lines.add(
-            new MouseWithIconClientTooltipComponent(Type.LEFT, getResourceRendering(asItem.getResource()), null)
-        ));
-        PlatformApi.INSTANCE.createResource(carried, true).ifPresent(asAlternative -> {
-            if (asAlternative.getResource() instanceof ItemResource) {
-                return;
-            }
-            lines.add(new MouseWithIconClientTooltipComponent(
-                Type.RIGHT,
-                getResourceRendering(asAlternative.getResource()),
-                null
-            ));
-        });
-        return lines;
-    }
-
-    public static <T> IconRenderer getResourceRendering(final T resource) {
-        return (graphics, x, y) -> PlatformApi.INSTANCE.getResourceRendering(resource).render(resource, graphics, x, y);
     }
 
     @FunctionalInterface